+++ unit_test/graph_test.cpp
@@ -76,24 +76,24 @@ TEST_F(TestInit, test_add_downstream_slow) {
 TEST_F(TestInit, test_get_nodes) {
   Graph* g = new Graph(nullptr);
   g->fast_nodes.resize(Graph::fast_node_max_num);
-  for (size_t i=0; i<Graph::fast_node_max_num; i++) {
+  for (uint64_t i=0; i<Graph::fast_node_max_num; i++) {
     std::shared_ptr<Node> shared_node;
     shared_node.reset((Node1*)(i+1));
     g->fast_nodes[i] = shared_node;
   }
-  for (size_t i=Graph::fast_node_max_num; i<200; i++) {
+  for (uint64_t i=Graph::fast_node_max_num; i<200; i++) {
     std::shared_ptr<Node> shared_node;
     shared_node.reset((Node1*)(i+1));
     g->nodes.push_back(shared_node);
   }
 
-  for (size_t i=0; i<200; i++) {
+  for (uint64_t i=0; i<200; i++) {
     EXPECT_EQ((void*)(g->get_node(i)), (void*)(i+1));
   }
 
   std::vector<Node*> required_nodes;
   g->get_nodes(required_nodes);
-  for (size_t i=0; i<200; i++) {
+  for (uint64_t i=0; i<200; i++) {
     EXPECT_EQ((void*)(required_nodes[i]), (void*)(i+1));
   }
 }
@@ -130,10 +130,10 @@ TEST_F(TestInit, test_create_edges) {
 
   std::vector<Node*> _nodes;
   g->node_num = 60;
-  for (size_t i=0; i<10; i++) {
+  for (uint64_t i=0; i<10; i++) {
     auto* t_ = new Node1();
     _nodes.push_back(t_);
-    size_t id = g->node_num;
+    uint64_t id = g->node_num;
     g->create_edges(t_, {t});
     EXPECT_EQ(t_->node_id, id);
     EXPECT_EQ(t_->wait_count, 1);
@@ -152,9 +152,9 @@ TEST_F(TestInit, test_create_edges) {
   EXPECT_EQ(g->nodes[5].get(), _nodes[8]);
   EXPECT_EQ(g->nodes[6].get(), _nodes[9]);
 
-  std::vector<size_t> vec_s;
+  std::vector<uint64_t> vec_s;
   bit_map_idx(t->fast_down_strams, 63, vec_s);
-  EXPECT_EQ(vec_s, std::vector<size_t>({60, 61, 62 }));
+  EXPECT_EQ(vec_s, std::vector<uint64_t>({60, 61, 62 }));
   EXPECT_EQ(t->down_streams, std::vector<Node*>({_nodes[3], _nodes[4], _nodes[5], _nodes[6], _nodes[7], _nodes[8], _nodes[9]}));
 }
 
@@ -201,7 +201,7 @@ TEST_F(TestInit, test_create_edges_finish) {
   EXPECT_EQ(g->fast_nodes[10].get(), t);
 
   std::vector<Node*> _nodes;
-  for (size_t i=0; i<10; i++) {
+  for (uint64_t i=0; i<10; i++) {
     auto* t_ = new Node1();
     _nodes.push_back(t_);
     if (i%2 == 0) {
@@ -218,7 +218,7 @@ TEST_F(TestInit, test_create_edges_finish) {
 
   Node1* t2 = new Node1(); 
   _nodes.clear();
-  for (size_t i=0; i<10; i++) {
+  for (uint64_t i=0; i<10; i++) {
     auto* t_ = new Node1();
     _nodes.push_back(t_);
     _nodes.push_back(t_);
