+++ unit_test/util_test.cpp
@@ -35,7 +35,7 @@ TEST_F(TestInit, test_set_bit) {
   EXPECT_FALSE(bit_map_get(bit_map, 3, 2));
   EXPECT_FALSE(bit_map_get(bit_map, 1, 2));
 
-  size_t bit_ = 20;
+  uint64_t bit_ = 20;
   bit_map = 0;
   bit_map_res = 0x1; bit_map_res <<= bit_;
   bit_map_set(bit_map, bit_, 30);
@@ -60,7 +60,7 @@ TEST_F(TestInit, test_set_bit) {
 
 TEST_F(TestInit, test_set) {
   long int bit_map = 0, bit_map_res = 0;
-  std::vector<size_t> vec_s;
+  std::vector<uint64_t> vec_s;
 
   bit_map_set(bit_map, 1, 0);
   EXPECT_EQ(bit_map_res, bit_map);
@@ -72,14 +72,14 @@ TEST_F(TestInit, test_set) {
   EXPECT_FALSE(bit_map_get(bit_map, 39, 63));
   EXPECT_FALSE(bit_map_get(bit_map, 32, 63));
   EXPECT_FALSE(bit_map_get(bit_map, 41, 63));
-  EXPECT_EQ(vec_s, std::vector<size_t>({40}));
+  EXPECT_EQ(vec_s, std::vector<uint64_t>({40}));
 
   bit_map_set(bit_map, 32, 63);
   bit_map_idx(bit_map, 63, vec_s);
   EXPECT_TRUE(bit_map_get(bit_map, 40, 63));
   EXPECT_TRUE(bit_map_get(bit_map, 32, 63));
   EXPECT_FALSE(bit_map_get(bit_map, 41, 63));
-  EXPECT_EQ(vec_s, std::vector<size_t>({32, 40}));
+  EXPECT_EQ(vec_s, std::vector<uint64_t>({32, 40}));
 
   bit_map_set(bit_map, 18, 63);
   bit_map_idx(bit_map, 63, vec_s);
@@ -87,7 +87,7 @@ TEST_F(TestInit, test_set) {
   EXPECT_TRUE(bit_map_get(bit_map, 32, 63));
   EXPECT_TRUE(bit_map_get(bit_map, 18, 63));
   EXPECT_FALSE(bit_map_get(bit_map, 41, 63));
-  EXPECT_EQ(vec_s, std::vector<size_t>({18, 32, 40}));
+  EXPECT_EQ(vec_s, std::vector<uint64_t>({18, 32, 40}));
 
   bit_map_set(bit_map, 62, 63);
   bit_map_idx(bit_map, 63, vec_s);
@@ -96,7 +96,7 @@ TEST_F(TestInit, test_set) {
   EXPECT_TRUE(bit_map_get(bit_map, 18, 63));
   EXPECT_TRUE(bit_map_get(bit_map, 62, 63));
   EXPECT_FALSE(bit_map_get(bit_map, 41, 63));
-  EXPECT_EQ(vec_s, std::vector<size_t>({18, 32, 40, 62}));
+  EXPECT_EQ(vec_s, std::vector<uint64_t>({18, 32, 40, 62}));
 
   bit_map_set(bit_map, 0, 63);
   bit_map_idx(bit_map, 63, vec_s);
@@ -106,7 +106,7 @@ TEST_F(TestInit, test_set) {
   EXPECT_TRUE(bit_map_get(bit_map, 62, 63));
   EXPECT_FALSE(bit_map_get(bit_map, 41, 63));
   EXPECT_FALSE(bit_map_get(bit_map, 63, 63));
-  EXPECT_EQ(vec_s, std::vector<size_t>({0, 18, 32, 40, 62}));
+  EXPECT_EQ(vec_s, std::vector<uint64_t>({0, 18, 32, 40, 62}));
 
   EXPECT_THROW(bit_map_idx(bit_map, 64, vec_s), std::runtime_error);
 }
